[lesson-1] create employee use case, refactor code

+++ internal/adapter/in_memory_db/position_db/find.go
@@ -2,7 +2,7 @@ package position_db
 
 import "warehouse_project/internal/domain/model"
 
-func (pr *PositionRepo) FindPosition(id int) (*model.Position, bool) {
+func (pr *PositionRepo) FindPosition(id int64) (*model.Position, bool) {
 	product, found := pr.positionsMap[id]
 	return product, found
 }
